requirements.txt
flask
psycopg2-binary
cloudinary
python-dotenv